+++ sdk/storage/azure-storage-blob/azure/storage/blob/_shared/uploads_async.py
@@ -6,10 +6,9 @@
 # pylint: disable=no-self-use
 
 import asyncio
+import threading
 from asyncio import Lock
 from itertools import islice
-import threading
-
 from math import ceil
 
 import six
@@ -17,11 +16,6 @@ import six
 from . import encode_base64, url_quote
 from .request_handlers import get_length
 from .response_handlers import return_response_headers
-from .encryption import (
-    GCMBlobEncryptionStream,
-    get_blob_encryptor_and_padder,
-    _ENCRYPTION_PROTOCOL_V1,
-    _ENCRYPTION_PROTOCOL_V2)
 from .uploads import SubStream
 
 
@@ -52,23 +46,8 @@ async def upload_data_chunks(
         chunk_size=None,
         max_concurrency=None,
         stream=None,
-        encryption_options=None,
         progress_hook=None,
         **kwargs):
-
-    if encryption_options:
-        # V1 uses an encryptor/padder to encrypt each chunk
-        if encryption_options['version'] == _ENCRYPTION_PROTOCOL_V1:
-            encryptor, padder = get_blob_encryptor_and_padder(
-                encryption_options.get('cek'),
-                encryption_options.get('vector'),
-                uploader_class is not PageBlobChunkUploader)
-            kwargs['encryptor'] = encryptor
-            kwargs['padder'] = padder
-
-        # V2 wraps the data stream with an encryption stream
-        elif encryption_options['version'] == _ENCRYPTION_PROTOCOL_V2:
-            stream = GCMBlobEncryptionStream(encryption_options.get('cek'), stream)
 
     parallel = max_concurrency > 1
     if parallel and 'modified_access_conditions' in kwargs:
