+++ sdk/storage/azure-storage-blob/azure/storage/blob/_upload_helpers.py
@@ -6,34 +6,36 @@
 # pylint: disable=no-self-use
 
 from io import SEEK_SET, UnsupportedOperation
-from typing import Optional, Union, Any, TypeVar, TYPE_CHECKING # pylint: disable=unused-import
+from typing import TypeVar, TYPE_CHECKING
 
 import six
 from azure.core.exceptions import ResourceExistsError, ResourceModifiedError, HttpResponseError
 
-from ._shared.response_handlers import (
-    process_storage_error,
-    return_response_headers)
+from ._shared.response_handlers import process_storage_error, return_response_headers
 from ._shared.models import StorageErrorCode
 from ._shared.uploads import (
     upload_data_chunks,
     upload_substream_blocks,
     BlockBlobChunkUploader,
     PageBlobChunkUploader,
-    AppendBlobChunkUploader)
-from ._shared.encryption import (
-    encrypt_blob,
-    get_adjusted_upload_size,
-    generate_blob_encryption_data,
-    _ENCRYPTION_PROTOCOL_V2)
+    AppendBlobChunkUploader
+)
 from ._generated.models import (
     BlockLookupList,
     AppendPositionAccessConditions,
     ModifiedAccessConditions,
 )
+from ._encryption import (
+    GCMBlobEncryptionStream,
+    encrypt_blob,
+    get_adjusted_upload_size,
+    get_blob_encryptor_and_padder,
+    generate_blob_encryption_data,
+    _ENCRYPTION_PROTOCOL_V1,
+    _ENCRYPTION_PROTOCOL_V2
+)
 
 if TYPE_CHECKING:
-    from datetime import datetime # pylint: disable=unused-import
     BlobLeaseClient = TypeVar("BlobLeaseClient")
 
 _LARGE_BLOB_UPLOAD_MAX_READ_BUFFER_SIZE = 4 * 1024 * 1024
@@ -131,17 +133,23 @@ def upload_block_blob(  # pylint: disable=too-many-locals
 
         if use_original_upload_path:
             total_size = length
-            if encryption_options.get('key'):
+            if encryption_options and encryption_options.get('key'):
                 cek, iv, encryption_data = generate_blob_encryption_data(
                     encryption_options['key'],
                     encryption_options['version'])
                 headers['x-ms-meta-encryptiondata'] = encryption_data
-                encryption_options['cek'] = cek
-                encryption_options['vector'] = iv
+
+                if encryption_options['version'] == _ENCRYPTION_PROTOCOL_V1:
+                    encryptor, padder = get_blob_encryptor_and_padder(cek, iv, True)
+                    kwargs['encryptor'] = encryptor
+                    kwargs['padder'] = padder
 
                 # Adjust total_size for encryption V2
                 if encryption_options['version'] == _ENCRYPTION_PROTOCOL_V2:
+                    # Adjust total_size for encryption V2
                     total_size = adjusted_count
+                    # V2 wraps the data stream with an encryption stream
+                    stream = GCMBlobEncryptionStream(cek, stream)
 
             block_ids = upload_data_chunks(
                 service=client,
@@ -151,7 +159,6 @@ def upload_block_blob(  # pylint: disable=too-many-locals
                 max_concurrency=max_concurrency,
                 stream=stream,
                 validate_content=validate_content,
-                encryption_options=encryption_options,
                 progress_hook=progress_hook,
                 headers=headers,
                 **kwargs
@@ -220,8 +227,12 @@ def upload_page_blob(
             except AttributeError:
                 tier = premium_page_blob_tier
 
-        if encryption_options and encryption_options.get('data'):
-            headers['x-ms-meta-encryptiondata'] = encryption_options['data']
+        if encryption_options and encryption_options.get('key'):
+            cek, iv, encryption_data = generate_blob_encryption_data(
+                encryption_options['key'],
+                encryption_options['version'])
+            headers['x-ms-meta-encryptiondata'] = encryption_data
+
         blob_tags_string = kwargs.pop('blob_tags_string', None)
         progress_hook = kwargs.pop('progress_hook', None)
 
@@ -238,6 +249,12 @@ def upload_page_blob(
         if length == 0:
             return response
 
+        if encryption_options and encryption_options.get('key'):
+            if encryption_options['version'] == _ENCRYPTION_PROTOCOL_V1:
+                encryptor, padder = get_blob_encryptor_and_padder(cek, iv, False)
+                kwargs['encryptor'] = encryptor
+                kwargs['padder'] = padder
+
         kwargs['modified_access_conditions'] = ModifiedAccessConditions(if_match=response['etag'])
         return upload_data_chunks(
             service=client,
@@ -247,7 +264,6 @@ def upload_page_blob(
             stream=stream,
             max_concurrency=max_concurrency,
             validate_content=validate_content,
-            encryption_options=encryption_options,
             progress_hook=progress_hook,
             headers=headers,
             **kwargs)
