+++ sdk/storage/azure-storage-blob/azure/storage/blob/aio/_upload_helpers.py
@@ -6,34 +6,36 @@
 # pylint: disable=no-self-use
 
 from io import SEEK_SET, UnsupportedOperation
-from typing import Optional, Union, Any, TypeVar, TYPE_CHECKING # pylint: disable=unused-import
+from typing import TypeVar, TYPE_CHECKING
 
 import six
 from azure.core.exceptions import ResourceModifiedError, HttpResponseError
 
-from .._shared.response_handlers import (
-    process_storage_error,
-    return_response_headers)
+from .._shared.response_handlers import process_storage_error, return_response_headers
 from .._shared.uploads_async import (
     upload_data_chunks,
     upload_substream_blocks,
     BlockBlobChunkUploader,
     PageBlobChunkUploader,
-    AppendBlobChunkUploader)
-from .._shared.encryption import (
-    encrypt_blob,
-    get_adjusted_upload_size,
-    generate_blob_encryption_data,
-    _ENCRYPTION_PROTOCOL_V2)
+    AppendBlobChunkUploader
+)
 from .._generated.models import (
     BlockLookupList,
     AppendPositionAccessConditions,
     ModifiedAccessConditions,
 )
+from .._encryption import (
+    GCMBlobEncryptionStream,
+    encrypt_blob,
+    get_adjusted_upload_size,
+    get_blob_encryptor_and_padder,
+    generate_blob_encryption_data,
+    _ENCRYPTION_PROTOCOL_V1,
+    _ENCRYPTION_PROTOCOL_V2
+)
 from .._upload_helpers import _convert_mod_error, _any_conditions
 
 if TYPE_CHECKING:
-    from datetime import datetime # pylint: disable=unused-import
     BlobLeaseClient = TypeVar("BlobLeaseClient")
 
 
@@ -105,17 +107,23 @@ async def upload_block_blob(  # pylint: disable=too-many-locals
 
         if use_original_upload_path:
             total_size = length
-            if encryption_options.get('key'):
+            if encryption_options and encryption_options.get('key'):
                 cek, iv, encryption_data = generate_blob_encryption_data(
                     encryption_options['key'],
                     encryption_options['version'])
                 headers['x-ms-meta-encryptiondata'] = encryption_data
-                encryption_options['cek'] = cek
-                encryption_options['vector'] = iv
+
+                if encryption_options['version'] == _ENCRYPTION_PROTOCOL_V1:
+                    encryptor, padder = get_blob_encryptor_and_padder(cek, iv, True)
+                    kwargs['encryptor'] = encryptor
+                    kwargs['padder'] = padder
 
                 # Adjust total_size for encryption V2
                 if encryption_options['version'] == _ENCRYPTION_PROTOCOL_V2:
+                    # Adjust total_size for encryption V2
                     total_size = adjusted_count
+                    # V2 wraps the data stream with an encryption stream
+                    stream = GCMBlobEncryptionStream(cek, stream)
 
             block_ids = await upload_data_chunks(
                 service=client,
@@ -194,8 +202,12 @@ async def upload_page_blob(
             except AttributeError:
                 tier = premium_page_blob_tier
 
-        if encryption_options and encryption_options.get('data'):
-            headers['x-ms-meta-encryptiondata'] = encryption_options['data']
+        if encryption_options and encryption_options.get('key'):
+            cek, iv, encryption_data = generate_blob_encryption_data(
+                encryption_options['key'],
+                encryption_options['version'])
+            headers['x-ms-meta-encryptiondata'] = encryption_data
+
         blob_tags_string = kwargs.pop('blob_tags_string', None)
         progress_hook = kwargs.pop('progress_hook', None)
 
@@ -211,6 +223,12 @@ async def upload_page_blob(
             **kwargs)
         if length == 0:
             return response
+
+        if encryption_options and encryption_options.get('key'):
+            if encryption_options['version'] == _ENCRYPTION_PROTOCOL_V1:
+                encryptor, padder = get_blob_encryptor_and_padder(cek, iv, False)
+                kwargs['encryptor'] = encryptor
+                kwargs['padder'] = padder
 
         kwargs['modified_access_conditions'] = ModifiedAccessConditions(if_match=response['etag'])
         return await upload_data_chunks(
