+++ sdk/storage/azure-storage-blob/tests/test_blob_encryption.py
@@ -3,15 +3,9 @@
 # Licensed under the MIT License. See License.txt in the project root for
 # license information.
 # --------------------------------------------------------------------------
+
 import tempfile
-
-import pytest
-
-import unittest
-from io import (
-    StringIO,
-    BytesIO,
-)
+from io import StringIO, BytesIO
 from json import loads
 from os import (
     urandom,
@@ -20,29 +14,26 @@ from os import (
     unlink
 )
 
+import pytest
 from azure.core.exceptions import HttpResponseError
-from azure.storage.blob._shared.encryption import (
+from azure.storage.blob import BlobServiceClient, BlobType
+from azure.storage.blob._blob_client import _ERROR_UNSUPPORTED_METHOD_FOR_ENCRYPTION
+from azure.storage.blob._encryption import (
     _dict_to_encryption_data,
     _validate_and_unwrap_cek,
     _generate_AES_CBC_cipher,
     _ERROR_OBJECT_INVALID,
 )
-from azure.storage.blob._blob_client import _ERROR_UNSUPPORTED_METHOD_FOR_ENCRYPTION
 from cryptography.hazmat.primitives.padding import PKCS7
-from devtools_testutils import ResourceGroupPreparer, StorageAccountPreparer
-from azure.storage.blob import (
-    BlobServiceClient,
-    ContainerClient,
-    BlobClient,
-    BlobType
-)
+
+from devtools_testutils.storage import StorageTestCase
 from encryption_test_helper import (
     KeyWrapper,
     KeyResolver,
     RSAKeyWrapper,
 )
 from settings.testcase import BlobPreparer
-from devtools_testutils.storage import StorageTestCase
+
 
 # ------------------------------------------------------------------------------
 TEST_CONTAINER_PREFIX = 'encryption_container'
@@ -51,6 +42,7 @@ TEST_BLOB_PREFIXES = {'BlockBlob': 'encryption_block_blob',
                       'AppendBlob': 'foo'}
 _ERROR_UNSUPPORTED_METHOD_FOR_ENCRYPTION = 'The require_encryption flag is set, but encryption is not supported' + \
                                            ' for this method.'
+# ------------------------------------------------------------------------------
 
 
 class StorageBlobEncryptionTest(StorageTestCase):
