+++ sdk/storage/azure-storage-file-share/azure/storage/fileshare/_shared/uploads_async.py
@@ -6,10 +6,9 @@
 # pylint: disable=no-self-use
 
 import asyncio
+import threading
 from asyncio import Lock
 from itertools import islice
-import threading
-
 from math import ceil
 
 import six
@@ -17,12 +16,7 @@ import six
 from . import encode_base64, url_quote
 from .request_handlers import get_length
 from .response_handlers import return_response_headers
-from .encryption import get_blob_encryptor_and_padder
-from .uploads import SubStream, IterStreamer  # pylint: disable=unused-import
-
-
-_LARGE_BLOB_UPLOAD_MAX_READ_BUFFER_SIZE = 4 * 1024 * 1024
-_ERROR_VALUE_SHOULD_BE_SEEKABLE_STREAM = '{0} should be a seekable file-like/io.IOBase type stream object.'
+from .uploads import SubStream
 
 
 async def _parallel_uploads(uploader, pending, running):
@@ -52,17 +46,8 @@ async def upload_data_chunks(
         chunk_size=None,
         max_concurrency=None,
         stream=None,
-        encryption_options=None,
         progress_hook=None,
         **kwargs):
-
-    if encryption_options:
-        encryptor, padder = get_blob_encryptor_and_padder(
-            encryption_options.get('cek'),
-            encryption_options.get('vector'),
-            uploader_class is not PageBlobChunkUploader)
-        kwargs['encryptor'] = encryptor
-        kwargs['padder'] = padder
 
     parallel = max_concurrency > 1
     if parallel and 'modified_access_conditions' in kwargs:
@@ -152,7 +137,6 @@ class _ChunkUploader(object):  # pylint: disable=too-many-instance-attributes
         self.parallel = parallel
 
         # Stream management
-        self.stream_start = stream.tell() if parallel else None
         self.stream_lock = threading.Lock() if parallel else None
 
         # Progress feedback
@@ -217,7 +201,7 @@ class _ChunkUploader(object):  # pylint: disable=too-many-instance-attributes
             self.progress_total += length
 
         if self.progress_hook:
-            self.progress_hook(self.progress_total, self.total_size)
+            await self.progress_hook(self.progress_total, self.total_size)
 
     async def _upload_chunk(self, chunk_offset, chunk_data):
         raise NotImplementedError("Must be implemented by child class.")
